warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@8691 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxManage/VBoxManage.cpp
@@ -5582,7 +5582,7 @@ static int handleControlVM(int argc, char *argv[],
 
             /* Process the command line. */
             int i;
-            for (i = 1 + 1; i < argc && cScancodes < RT_ELEMENTS(alScancodes); i++, cScancodes++)
+            for (i = 1 + 1; i < argc && cScancodes < (int)RT_ELEMENTS(alScancodes); i++, cScancodes++)
             {
                 if (   isxdigit (argv[i][0])
                     && isxdigit (argv[i][1])
